old_project_to_be_base_on/write_your_code_base_on_me/requirements.txt
Flask==3.0.2
flask_mail==0.9.1
password_strength==0.0.3.post2
pymssql==2.2.11
python-dotenv==1.0.1
